Split table header into separate components
